+++ rustsat/src/encodings/card/simulators.rs
@@ -171,7 +171,7 @@ where
         self.card_enc.n_clauses()
     }
 
-    fn n_vars(&self) -> usize {
+    fn n_vars(&self) -> u32 {
         self.card_enc.n_vars()
     }
 }
@@ -311,7 +311,7 @@ where
         self.ub_enc.n_clauses() + self.lb_enc.n_clauses()
     }
 
-    fn n_vars(&self) -> usize {
+    fn n_vars(&self) -> u32 {
         self.ub_enc.n_vars() + self.lb_enc.n_vars()
     }
 }
